CALID/requirements.txt
matplotlib==3.8.0
numpy==1.25.0
networkx==3.1
opencv-python-headless==4.8.1.78
pandas==2.0.3
Pillow==10.0.0
pytorch-lightning==1.5.10
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.1
semantic-version==2.10.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0

